requirements.txt
djangorestframework
django-cors-headers
pillow
djoser
pytils
django-ckeditor
segno
faker
django-random-queryset